cheri: Enable virtio HW RNG support in kernel config
Enable support for hardware random number generators in particular
the QEMU virtio-rng one.

+++ drivers/char/hw_random/atmel-rng.c
@@ -136,7 +136,7 @@ static int atmel_trng_probe(struct platform_device *pdev)
 	trng->has_half_rate = data->has_half_rate;
 	trng->rng.name = pdev->name;
 	trng->rng.read = atmel_trng_read;
-	trng->rng.priv = (unsigned long)&pdev->dev;
+	trng->rng.priv = (uintptr_t)&pdev->dev;
 	platform_set_drvdata(pdev, trng);
 
 #ifndef CONFIG_PM
